feat & rename : made src & constants folder - src folder for organize folders - constants folder for constant variables
