+++ src/app/addArticle/page.tsx
@@ -36,20 +36,14 @@ export default function AddAritcle() {
     }
   }
 
+  const handleChangeTitle = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => setTitle(value)
+
+  const handleChangeContent = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => setContent(value)
+
   return (
     <form onSubmit={handleSubmit}>
-      <input
-        onChange={({ target: { value } }: ChangeEvent<HTMLInputElement>) => setTitle(value)}
-        value={title}
-        type="text"
-        placeholder="Text title..."
-      />
-      <input
-        onChange={({ target: { value } }: ChangeEvent<HTMLInputElement>) => setContent(value)}
-        value={content}
-        type="text"
-        placeholder="Text content..."
-      />
+      <input onChange={handleChangeTitle} value={title} type="text" placeholder="Text title..." />
+      <input onChange={handleChangeContent} value={content} type="text" placeholder="Text content..." />
       <button type="submit">Publish</button>
     </form>
   )
