+++ src/constants/common.ts
@@ -1,0 +1,1 @@
+const DEFINED_URL = process.env.NODE_ENV === 'development' ? process.env.DEV_URL : process.env.PROD_URL
